모든 셀 구성 메서드 수정 및 타입 프로퍼티 identifier 선언, 초기화 / TravelModel grade 타입 변경(Float -> Double) / AdCell에서 layoutSubviews 이용해 구분선 지우기 / CityCell에서 prepareForReuse 이용해 특정 셀에서 지운 구분선 다시 그리기

+++ TableView_Travel/Model/TravelModel.swift
@@ -10,7 +10,7 @@ struct Travel {
     let title: String
     let description: String?
     let travel_image: String?
-    let grade: Float?
+    let grade: Double?
     let save: Int?
     var like: Bool?
     let ad: Bool
